Move some files to where they fit better. Although the data directory is now empty, I intend to put some stuff in there in due course.
